// @types/generated/contentful.d.ts
export interface I3ZPkEj1KqeSn4QdsdnNko3Fields {
  /** Personalized Hero Name */
  name: string;

  /** Hero Variants */
  unfrmOptP13nList: IHero[];
}

export interface I3ZPkEj1KqeSn4QdsdnNko3
  extends Entry<I3ZPkEj1KqeSn4QdsdnNko3Fields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "3zPkEj1KqeSn4QdsdnNKO3";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IAbTestFields {
  /** List Name */
  name: string;

  /** A/B Test */
  unfrmAbTest: Record<string, any>;

  /** Variants */
  variants: IAbTestVariant[];

  /** Personalization Criteria */
  unfrmOptPersonalizationCriteria?: Record<string, any> | undefined;
}

export interface IAbTest extends Entry<IAbTestFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "ABTest";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IAbTestVariantFields {
  /** Identifier */
  identifier?: string | undefined;

  /** Distribution */
  distribution?: number | undefined;

  /** Component */
  component?: Entry<{ [fieldId: string]: unknown }> | undefined;
}

export interface IAbTestVariant extends Entry<IAbTestVariantFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "abTestVariant";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface ICta extends Entry<ICtaFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "cta";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IHeroFields {
  /** Title */
  title: string;

  /** Description */
  description: string;

  /** Button Text */
  buttonText?: string | undefined;

  /** Button Link Slug */
  buttonLinkSlug?: string | undefined;

  /** Image */
  image?: Asset | undefined;

  /** Personalization Criteria */
  unfrmOptPersonalizationCriteria?: Record<string, any> | undefined;
}

export interface IHero extends Entry<IHeroFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "hero";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IPageFields {
  /** Title */
  title?: string | undefined;

  /** Slug */
  slug?: string | undefined;

  /** Components */
  components?:
    | (
        | ICta
        | IHero
        | I3ZPkEj1KqeSn4QdsdnNko3
        | IRegistrationForm
        | ITalksList
        | IWhyAttend
        | IAbTest
      )[]
    | undefined;

  /** Enrichment Tags */
  unfrmOptEnrichmentTag?: Record<string, any> | undefined;
}

export interface IPage extends Entry<IPageFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "page";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}
// ...
export interface IRegistrationFormFields {
  /** Heading */
  heading?: string | undefined;

  /** ButtonText */
  buttonText?: string | undefined;

  /** Registered Text */
  registeredText?: string | undefined;
}

export interface IRegistrationForm extends Entry<IRegistrationFormFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "registrationForm";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ITalkFields {
  /** Title */
  title: string;

  /** Description */
  description: Document;

  /** Audience */
  audience?: ("Developers" | "Marketers")[] | undefined;

  /** Personalization Criteria */
  unfrmOptPersonalizationCriteria?: Record<string, any> | undefined;
}

export interface ITalk extends Entry<ITalkFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "talk";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface ITalksListFields {
  /** Title */
  title: string;

  /** NumberToShow */
  count?: number | undefined;
}

export interface ITalksList extends Entry<ITalksListFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "talksList";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}

export interface IWhyAttendFields {
  /** Title */
  title?: string | undefined;

  /** Description */
  description?: string | undefined;

  /** Image */
  image?: Asset | undefined;

  /** Align to */
  alignToTheLeft?: boolean | undefined;
}

export interface IWhyAttend extends Entry<IWhyAttendFields> {
  sys: {
    id: string;
    type: string;
    createdAt: string;
    updatedAt: string;
    locale: string;
    contentType: {
      sys: {
        id: "whyAttend";
        linkType: "ContentType";
        type: "Link";
      };
    };
  };
}